Reestructura de todo el proyecto en formato API REST
