+++ src/main/java/mx/ipn/escom/pizarron/crud/application/CrudConfig.java
@@ -1,9 +1,7 @@
 package mx.ipn.escom.pizarron.crud.application;
 
-import mx.ipn.escom.pizarron.crud.domain.impl.CatalogServiceImpl;
-import mx.ipn.escom.pizarron.crud.domain.impl.UserServiceImpl;
-import mx.ipn.escom.pizarron.crud.domain.service.CatalogService;
-import mx.ipn.escom.pizarron.crud.domain.service.UserService;
+import mx.ipn.escom.pizarron.crud.domain.impl.*;
+import mx.ipn.escom.pizarron.crud.domain.service.*;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -11,12 +9,32 @@ import org.springframework.context.annotation.Configuration;
 public class CrudConfig {
 
     @Bean
-    public CatalogService catalogServicePort() {
-        return new CatalogServiceImpl();
+    public ClassService classServicePort() {
+        return new ClassServiceImpl();
     }
 
     @Bean
-    public UserService userServicePort() {
-        return new UserServiceImpl();
-    }
+    public GroupService groupServicePort() { return new GroupServiceImpl(); }
+
+    @Bean
+    public GroupSubjectService groupSubjectService() { return new GroupSubjectServiceImpl(); }
+
+    @Bean
+    public PermissionService permissionService () { return new PermissionServiceImpl(); }
+
+    @Bean
+    public SchoolCycleService schoolCycleService() { return new SchoolCycleServiceImpl(); }
+
+    @Bean
+    public SubjectService subjectService () { return new SubjectServiceImpl() ; }
+
+    @Bean
+    public UserRolService userRolService () { return new UserRolServiceImpl() ; }
+
+    @Bean
+    public UserService userServicePort() { return new UserServiceImpl(); }
+
+    @Bean
+    public UserStatusService userStatusService() { return new UserStatusServiceImpl(); }
+
 }
