+++ src/main/java/mx/ipn/escom/pizarron/crud/domain/impl/ClassServiceImpl.java
@@ -1,0 +1,11 @@
+package mx.ipn.escom.pizarron.crud.domain.impl;
+
+import lombok.extern.log4j.Log4j2;
+import mx.ipn.escom.pizarron.crud.domain.service.ClassService;
+import org.jvnet.hk2.annotations.Service;
+
+@Service
+@Log4j2
+public class ClassServiceImpl implements ClassService {
+
+}
