+++ src/main/java/mx/ipn/escom/pizarron/crud/domain/service/ClassService.java
@@ -1,0 +1,4 @@
+package mx.ipn.escom.pizarron.crud.domain.service;
+
+public interface ClassService {
+}
