+++ src/main/java/mx/ipn/escom/pizarron/crud/port/controller/ClassController.java
@@ -1,0 +1,11 @@
+package mx.ipn.escom.pizarron.crud.port.controller;
+
+import lombok.extern.log4j.Log4j2;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/class")
+@Log4j2
+public class ClassController {
+}
