+++ src/main/java/mx/ipn/escom/pizarron/crud/port/controller/UserController.java
@@ -7,7 +7,6 @@ import mx.ipn.escom.pizarron.crud.adapter.dto.request.simple.DataRequest;
 import mx.ipn.escom.pizarron.crud.adapter.dto.response.UserResponseDto;
 import mx.ipn.escom.pizarron.crud.adapter.dto.response.simple.DataResponse;
 import mx.ipn.escom.pizarron.crud.adapter.entity.UserEntity;
-import mx.ipn.escom.pizarron.crud.adapter.entity.UserPermissionEntity;
 import mx.ipn.escom.pizarron.crud.domain.service.UserService;
 import mx.ipn.escom.pizarron.crud.util.ExceptionMessage;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,10 +15,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/users")
 @Log4j2
 public class UserController {
 
@@ -28,32 +25,30 @@ public class UserController {
 
     @PostMapping
     public ResponseEntity<DataResponse<UserResponseDto>> createUser(@RequestBody UserRequestDto userRequestDto) {
-        UserResponseDto userEntity = null;
         try{
-            userEntity = userService.createUser(userRequestDto);
+            return new ResponseEntity<>(new DataResponse<>(userService.createUser(userRequestDto)), HttpStatus.OK);
         }catch (Exception businessException){
             return new ResponseEntity<>(new DataResponse<>(null, false, HttpStatus.BAD_REQUEST.value(),
                     ExceptionMessage.setMessage(businessException.getMessage())), HttpStatus.BAD_REQUEST);
         }
-        return new ResponseEntity<>(new DataResponse<>(userEntity), HttpStatus.OK);
     }
 
-    @PutMapping
-    public ResponseEntity<DataResponse<UserResponseDto>> updateUser(@RequestBody UserRequestDto userRequestDto) {
-        UserResponseDto userEntity = null;
+    @PutMapping("/{key}")
+    public ResponseEntity<DataResponse<UserResponseDto>> updateUser(@PathVariable(value="key") String key,
+                                                                    @RequestBody UserRequestDto userRequestDto) {
         try{
-            userEntity = userService.updateUser(userRequestDto);
+            return new ResponseEntity<>(new DataResponse<>(userService.updateUser(key, userRequestDto)), HttpStatus.OK);
         }catch (Exception businessException){
             return new ResponseEntity<>(new DataResponse<>(null, false, HttpStatus.BAD_REQUEST.value(),
                     ExceptionMessage.setMessage(businessException.getMessage())), HttpStatus.BAD_REQUEST);
         }
-        return new ResponseEntity<>(new DataResponse<>(userEntity), HttpStatus.OK);
     }
 
-    @GetMapping
-    public ResponseEntity<DataResponse<UserResponseDto>> getUser(@RequestBody UserRequestDto userRequestDto) {
+    @GetMapping("/{key}")
+    public ResponseEntity<DataResponse<UserResponseDto>> getUser(@PathVariable(value="key") String key,
+                                                                 @RequestBody UserRequestDto userRequestDto) {
         try{
-            UserResponseDto userEntity = userService.getUser(userRequestDto);
+            UserResponseDto userEntity = userService.getUser(key, userRequestDto);
             return new ResponseEntity<>(new DataResponse<>(userEntity), HttpStatus.OK);
         }catch (Exception businessException){
             return new ResponseEntity<>(new DataResponse<>(null, false, HttpStatus.BAD_REQUEST.value(),
@@ -61,9 +56,8 @@ public class UserController {
         }
     }
 
-    @GetMapping("/list")
+    @GetMapping
     public ResponseEntity<DataResponse<Page<UserEntity>>> getUserList(@RequestBody DataRequest<UserRequestDto> userRequestDto) {
-        log.info(userRequestDto);
         try{
             Page<UserEntity> userEntity = userService.getUserList(userRequestDto);
             return new ResponseEntity<>(new DataResponse<>(userEntity), HttpStatus.OK);
@@ -73,16 +67,14 @@ public class UserController {
         }
     }
 
-    @PutMapping("/permissions")
-    public ResponseEntity<DataResponse<UserResponseDto>> updatePermissions(@RequestBody UserPermissionRequestDto userPermissionRequestDto) {
-        log.info(userPermissionRequestDto);
-        UserResponseDto userPermission = null;
+    @PutMapping("/permissions/{key}")
+    public ResponseEntity<DataResponse<UserResponseDto>> updatePermissions(@PathVariable(value="key") String key,
+                                                                           @RequestBody UserPermissionRequestDto userPermissionRequestDto) {
         try{
-            userPermission = userService.updatePermissions(userPermissionRequestDto);
+            return new ResponseEntity<>(new DataResponse<>(userService.updatePermissions(key, userPermissionRequestDto)), HttpStatus.OK);
         }catch (Exception businessException){
             return new ResponseEntity<>(new DataResponse<>(null, false, HttpStatus.BAD_REQUEST.value(),
                     ExceptionMessage.setMessage(businessException.getMessage())), HttpStatus.BAD_REQUEST);
         }
-        return new ResponseEntity<>(new DataResponse<>(userPermission), HttpStatus.OK);
     }
 }
