+++ src/main/java/mx/ipn/escom/pizarron/crud/port/controller/UserStatusController.java
@@ -1,0 +1,35 @@
+package mx.ipn.escom.pizarron.crud.port.controller;
+
+import lombok.extern.log4j.Log4j2;
+import mx.ipn.escom.pizarron.crud.adapter.dto.request.simple.SimpleRequest;
+import mx.ipn.escom.pizarron.crud.adapter.dto.response.simple.DataResponse;
+import mx.ipn.escom.pizarron.crud.adapter.entity.UserStatusEntity;
+import mx.ipn.escom.pizarron.crud.domain.service.UserStatusService;
+import mx.ipn.escom.pizarron.crud.util.ExceptionMessage;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/user-statuses")
+@Log4j2
+public class UserStatusController {
+
+    @Autowired
+    UserStatusService userStatusService;
+
+    @GetMapping("/user-statuses")
+    public ResponseEntity<DataResponse<Page<UserStatusEntity>>> getUserStatusList(@RequestBody SimpleRequest simpleRequest) {
+        try{
+            return new ResponseEntity<>(new DataResponse<>(userStatusService.getUserStatusList(simpleRequest)), HttpStatus.OK);
+        }catch (Exception businessException){
+            return new ResponseEntity<>(new DataResponse<>(null, false, HttpStatus.BAD_REQUEST.value(),
+                    ExceptionMessage.setMessage(businessException.getMessage())), HttpStatus.BAD_REQUEST);
+        }
+    }
+}
